memory/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
backoff==2.2.1
certifi==2025.8.3
charset-normalizer==3.4.3
distro==1.9.0
dotenv==0.9.9
grpcio==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
mem0ai==0.1.116
numpy==2.3.2
openai==1.101.0
portalocker==3.2.0
posthog==6.6.1
protobuf==5.29.5
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
qdrant-client==1.15.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
